Split C code into more files
