+++ src/eip4844/api.h
@@ -16,7 +16,7 @@
 
 #pragma once
 
-#include "common.h"
+#include "common/common.h"
 
 #ifdef __cplusplus
 extern "C" {
