+++ src/eip7594/constants.h
@@ -16,13 +16,8 @@
 
 #pragma once
 
-#include "common.h"
-
-#include <stdint.h>
-
-#ifdef __cplusplus
-extern "C" {
-#endif
+#include "blst.h"
+#include "common/common.h"
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 // Macros
@@ -38,41 +33,39 @@ extern "C" {
 #define BYTES_PER_CELL (FIELD_ELEMENTS_PER_CELL * BYTES_PER_FIELD_ELEMENT)
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
-// Types
+// Constants
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
-/** A single cell for a blob. */
-typedef struct {
-    uint8_t bytes[BYTES_PER_CELL];
-} Cell;
+/** The domain separator for verify_cell_kzg_proof_batch's random challenge. */
+static const char *RANDOM_CHALLENGE_DOMAIN_VERIFY_CELL_KZG_PROOF_BATCH = "RCKZGCBATCH__V1_";
 
-////////////////////////////////////////////////////////////////////////////////////////////////////
-// Public Functions
-////////////////////////////////////////////////////////////////////////////////////////////////////
+/**
+ * The coset shift factor for the cell recovery code.
+ *
+ *   fr_t a;
+ *   fr_from_uint64(&a, 7);
+ *   for (size_t i = 0; i < 4; i++)
+ *       printf("%#018llxL,\n", a.l[i]);
+ */
+static const fr_t RECOVERY_SHIFT_FACTOR = {
+    0x0000000efffffff1L,
+    0x17e363d300189c0fL,
+    0xff9c57876f8457b0L,
+    0x351332208fc5a8c4L,
+};
 
-C_KZG_RET compute_cells_and_kzg_proofs(
-    Cell *cells, KZGProof *proofs, const Blob *blob, const KZGSettings *s
-);
-
-C_KZG_RET recover_cells_and_kzg_proofs(
-    Cell *recovered_cells,
-    KZGProof *recovered_proofs,
-    const uint64_t *cell_indices,
-    const Cell *cells,
-    size_t num_cells,
-    const KZGSettings *s
-);
-
-C_KZG_RET verify_cell_kzg_proof_batch(
-    bool *ok,
-    const Bytes48 *commitments_bytes,
-    const uint64_t *cell_indices,
-    const Cell *cells,
-    const Bytes48 *proofs_bytes,
-    size_t num_cells,
-    const KZGSettings *s
-);
-
-#ifdef __cplusplus
-}
-#endif
+/**
+ * The inverse of RECOVERY_SHIFT_FACTOR.
+ *
+ *   fr_t a;
+ *   fr_from_uint64(&a, 7);
+ *   fr_div(&a, &FR_ONE, &a);
+ *   for (size_t i = 0; i < 4; i++)
+ *       printf("%#018llxL,\n", a.l[i]);
+ */
+static const fr_t INV_RECOVERY_SHIFT_FACTOR = {
+    0xdb6db6dadb6db6dcL,
+    0xe6b5824adb6cc6daL,
+    0xf8b356e005810db9L,
+    0x66d0f1e660ec4796L,
+};
