+++ src/eip7594/types.h
@@ -13,35 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 #pragma once
 
-#include "common.h"
+#include "constants.h"
 
-#include <stdio.h>
-
-#ifdef __cplusplus
-extern "C" {
-#endif
+#include <inttypes.h>
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
-// Public Functions
+// Types
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
-C_KZG_RET load_trusted_setup(
-    KZGSettings *out,
-    const uint8_t *g1_monomial_bytes,
-    size_t num_g1_monomial_bytes,
-    const uint8_t *g1_lagrange_bytes,
-    size_t num_g1_lagrange_bytes,
-    const uint8_t *g2_monomial_bytes,
-    size_t num_g2_monomial_bytes,
-    size_t precompute
-);
-
-C_KZG_RET load_trusted_setup_file(KZGSettings *out, FILE *in, size_t precompute);
-
-void free_trusted_setup(KZGSettings *s);
-
-#ifdef __cplusplus
-}
-#endif
+/** A single cell for a blob. */
+typedef struct {
+    uint8_t bytes[BYTES_PER_CELL];
+} Cell;
